+++ AspNetCore/Domain/Models/LegalInformation.cs
@@ -1,0 +1,11 @@
+using Domain.Attributes;
+
+namespace Domain.Models
+{
+    [TypeSharp]
+    public class LegalInformationCreateDto
+    {
+        public string RegistrationNumber { get; set; }
+        public string? VatNumber { get; set; }
+    }
+}